+++ lib/concierge/cache/storage.rb
@@ -22,7 +22,7 @@ class Concierge::Cache
   class Storage
 
     def read(key)
-      EntryRepository.by_key(key)
+      database.with_repository { |repository| repository.by_key(key) }
     end
 
     def write(key, value)
@@ -30,10 +30,10 @@ class Concierge::Cache
 
       if entry
         entry.value = value
-        EntryRepository.update(entry)
+        database.update(entry)
       else
         entry = Entry.new(key: key, value: value, updated_at: Time.now)
-        EntryRepository.create(entry)
+        create_or_update(entry)
       end
 
       entry
@@ -41,7 +41,31 @@ class Concierge::Cache
 
     def delete(key)
       entry = read(key)
-      EntryRepository.delete(entry) if entry
+      database.delete(entry) if entry
+    end
+
+    private
+
+      # under some very specific situations, by the time the cache is being written,
+      # a different process executed a query for the same cache key and wrote it to
+      # the database, causing the creation call to fail with a unique constraint
+      # violation error. In technical terms, this is a time-of-check, time-of-use
+      # issue.
+      #
+      # This method works around such scenario by falling back to a standard update
+      # in case the creation of the record cause the error described above.
+      def create_or_update(entry)
+        EntryRepository.create(entry)
+      rescue Hanami::Model::UniqueConstraintViolationError
+        entry = read(entry.key)
+        database.update(entry)
+      rescue Hanami::Model::Error => err
+        database.create(entry)
+      end
+
+
+    def database
+      @database ||= Concierge::OptionalDatabaseAccess.new(EntryRepository)
     end
 
   end