+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This file summarises the most important changes that went live on each release
 of Concierge. Please check the Wiki entry on the release process to understand
 how this file is formatted and how the process works.
 
+## [0.1.4] - 2016-04-21
+- make Concierge operations resilient to database failures (introduces Emergency Log).
+- time-of-check, time-of-use issue on Concierge's cache. Fixes [Rollbar #8](https://rollbar.com/Roomorama/Concierge/items/8/).
+- valid name for customer with non-latin letters for JTB client
+
 ## [0.1.3] - 2016-04-07
 ### Changed
 - fixed a bug on JTB's response parser when a unit is not available (no rate plans returned.)